Pipe FUSE implementation: implemented rename method, refactored code, added build script
