+++ pipe-cli/mount/requirements.txt
@@ -1,0 +1,4 @@
+fusepy==3.0.1
+easywebdav==1.2.0
+dis3==0.1.3
+altgraph==0.16.1
